refactor: infrastructure 구조로 수정
